Partition reachable sets along automaton states instead of propositions
We do not care about the individual propositions anymore and can thus merge reachable sets with different propositions.
The important part is that we only merge reachable sets that are equivalent in terms of satisfaction of the traffic rules, i.e., those that are labeled with the same automaton states.