Track propositions outside of reach node
Accessing the private field of the reach interface in kripke.py and visualization.py is not optimal.
However, the situation before (assuming that the reach interface returns SemanticReachNode) was not better.
What we would really need is a reach interface tailored to the semantic part that allows to access propositions.